Create cleanClubSummarize.py
This program takes an excel workbook as input. The excel workbook contains sheet names as threat actors with column A populated with the TTP IDs, and column B with the sources.  It cleans the data by eliminating the redundant entries of TTP IDs for each actor, creates an ATT&CK Navigator .json file that combines the TTPs used by each actor, and produces a summary sheet for easily understanding the frequency of a specific TTP ID used by threat actors.